+++ src/test/java/de/is24/deadcode4j/plugin/IT_PuttingItAllTogether.java
@@ -36,7 +36,7 @@ public class IT_PuttingItAllTogether extends AbstractMojoTestCase {
     public void test() throws Exception {
         findDeadCodeMojo.execute();
 
-        verify(logMock).info("Analyzed 17 class(es).");
+        verify(logMock).info("Analyzed 18 class(es).");
         verify(logMock).info("Ignoring 1 class(es) which seem(s) to be unused.");
         verify(logMock).warn("Class [com.acme.Foo] should be ignored, but is not dead. You should remove the configuration entry.");
         verify(logMock).warn("Found 1 unused class(es):");